moved common files to folder
